client/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
pigpio==1.78
requests==2.32.3
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
urllib3==2.2.3
